+++ lib/polars/io/ipc.rb
@@ -82,6 +82,82 @@ module Polars
       Utils.wrap_df(rbdf)
     end
 
+    # Read into a DataFrame from Arrow IPC record batch stream.
+    #
+    # See "Streaming format" on https://arrow.apache.org/docs/python/ipc.html.
+    #
+    # @param source [Object]
+    #   Path to a file or a file-like object (by "file-like object" we refer to objects
+    #   that have a `read()` method, such as a file handler like the builtin `open`
+    #   function, or a `BytesIO` instance). If `fsspec` is installed, it will be used
+    #   to open remote files.
+    # @param columns [Array]
+    #   Columns to select. Accepts a list of column indices (starting at zero) or a list
+    #   of column names.
+    # @param n_rows [Integer]
+    #   Stop reading from IPC stream after reading `n_rows`.
+    # @param storage_options [Hash]
+    #   Extra options that make sense for a particular storage connection.
+    # @param row_index_name [String]
+    #   Insert a row index column with the given name into the DataFrame as the first
+    #   column. If set to `nil` (default), no row index column is created.
+    # @param row_index_offset [Integer]
+    #   Start the row index at this offset. Cannot be negative.
+    #   Only used if `row_index_name` is set.
+    # @param rechunk [Boolean]
+    #   Make sure that all data is contiguous.
+    #
+    # @return [DataFrame]
+    def read_ipc_stream(
+      source,
+      columns: nil,
+      n_rows: nil,
+      storage_options: nil,
+      row_index_name: nil,
+      row_index_offset: 0,
+      rechunk: true
+    )
+      storage_options ||= {}
+      _prepare_file_arg(source, **storage_options) do |data|
+        _read_ipc_stream_impl(
+          data,
+          columns: columns,
+          n_rows: n_rows,
+          row_index_name: row_index_name,
+          row_index_offset: row_index_offset,
+          rechunk: rechunk
+        )
+      end
+    end
+
+    # @private
+    def _read_ipc_stream_impl(
+      source,
+      columns: nil,
+      n_rows: nil,
+      row_index_name: nil,
+      row_index_offset: 0,
+      rechunk: true
+    )
+      if Utils.pathlike?(source)
+        source = Utils.normalize_filepath(source)
+      end
+      if columns.is_a?(String)
+        columns = [columns]
+      end
+
+      projection, columns = Utils.handle_projection_columns(columns)
+      pydf = RbDataFrame.read_ipc_stream(
+        source,
+        columns,
+        projection,
+        n_rows,
+        Utils._prepare_row_count_args(row_index_name, row_index_offset),
+        rechunk
+      )
+      Utils.wrap_df(pydf)
+    end
+
     # Get a schema of the IPC file without reading data.
     #
     # @param source [Object]